+++ extensions/smartthings-connector/.eslintrc.js
@@ -1,4 +1,7 @@
 module.exports = {
     root: true,
-    extends: ["@raycast"]
+    extends: ["@raycast"],
+    rules: {
+        "@typescript-eslint/no-explicit-any": "off"
+    }
 };